requirements.txt
astroid==1.5.3
-e git+https://github.com/django/django.git@7e3f8e05eb56cbd01f50b811c551a7a79b250ef7#egg=Django
django-cors-headers==2.1.0
django-filter==1.0.4
djangorestframework==3.6.4
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
mysqlclient==1.3.12
pytz==2017.2
six==1.11.0
wrapt==1.10.11
